chore(org): collocate source files & tests
