+++ src/__tests__/index.test.ts
@@ -1,22 +1,22 @@
-import authorize from '../api/authorize';
-import onboard from '../api/onboard';
-import logout from '../api/logout';
-import openService from '../api/open-service';
-import requestLoginLink from '../api/request-login-link';
-import exchangeToken from '../api/exchange-token';
-import tokenInfo from '../api/token-info';
+import authorize from '../api/authorize/authorize';
+import onboard from '../api/onboard/onboard';
+import logout from '../api/logout/logout';
+import openService from '../api/open-service/open-service';
+import requestLoginLink from '../api/request-login-link/request-login-link';
+import exchangeToken from '../api/exchange-token/exchange-token';
+import tokenInfo from '../api/token-info/token-info';
 import mtLinkSdk, { Mode, MtLinkSdk } from '..';
 
 import packageJson from '../../package.json';
 import expectUrlToMatchWithPKCE from './helper/expect-url-to-match';
 
-vi.mock('../api/authorize');
-vi.mock('../api/onboard');
-vi.mock('../api/logout');
-vi.mock('../api/open-service');
-vi.mock('../api/request-login-link');
-vi.mock('../api/exchange-token');
-vi.mock('../api/token-info');
+vi.mock('../api/authorize/authorize');
+vi.mock('../api/onboard/onboard');
+vi.mock('../api/logout/logout');
+vi.mock('../api/open-service/open-service');
+vi.mock('../api/request-login-link/request-login-link');
+vi.mock('../api/exchange-token/exchange-token');
+vi.mock('../api/token-info/token-info');
 
 describe('index', () => {
   test('MtLinkSdk', async () => {
