+++ src/api/authorize/authorize-url.test.ts
@@ -1,24 +1,21 @@
 import { MY_ACCOUNT_DOMAINS } from '../../server-paths';
-import { AuthorizeOptions, MtLinkSdk } from '../..';
-import authorize from '../authorize';
+import { MtLinkSdk } from '../..';
+import authorizeUrl from './authorize-url';
 import { generateConfigs } from '../../helper';
-import * as helper from '../../helper';
 import storage from '../../storage';
 import expectUrlToMatchWithPKCE from '../../__tests__/helper/expect-url-to-match';
 
 vi.mock('../../storage');
 
 describe('api', () => {
-  describe('authorize', () => {
-    const open = (window.open = vi.fn());
-
+  describe('authorize-url', () => {
     const mockedStorage = vi.mocked(storage);
 
     const clientId = 'clientId';
     const redirectUri = 'redirectUri';
 
     test('without calling init', async () => {
-      await expect(authorize(new MtLinkSdk().storedOptions)).rejects.toThrow(
+      await expect(authorizeUrl(new MtLinkSdk().storedOptions)).rejects.toThrow(
         '[mt-link-sdk] Make sure to call `init` before calling `authorizeUrl/authorize`.'
       );
     });
@@ -27,14 +24,13 @@ describe('api', () => {
       const mtLinkSdk = new MtLinkSdk();
       mtLinkSdk.init(clientId);
 
-      await expect(authorize(mtLinkSdk.storedOptions)).rejects.toThrow(
+      await expect(authorizeUrl(mtLinkSdk.storedOptions)).rejects.toThrow(
         '[mt-link-sdk] Missing option `redirectUri` in `authorizeUrl/authorize`, make sure to pass one via `authorizeUrl/authorize` options or `init` options.'
       );
     });
 
     test('method call without options use default init value', async () => {
       mockedStorage.set.mockClear();
-      open.mockClear();
 
       const country = 'JP';
       const scopes = 'points_read';
@@ -51,11 +47,8 @@ describe('api', () => {
         samlSubjectId
       });
 
-      await authorize(mtLinkSdk.storedOptions);
+      const url = await authorizeUrl(mtLinkSdk.storedOptions);
 
-      expect(open).toHaveBeenCalledTimes(1);
-      expect(open).toHaveBeenCalledWith(expect.any(String), '_self', 'noreferrer');
-      const url = open.mock.calls[0][0];
       const query = {
         client_id: clientId,
         cobrand_client_id: cobrandClientId,
@@ -72,7 +65,6 @@ describe('api', () => {
 
     test('with options', async () => {
       mockedStorage.set.mockClear();
-      open.mockClear();
 
       const state = 'state';
       const country = 'JP';
@@ -82,15 +74,12 @@ describe('api', () => {
       const mtLinkSdk = new MtLinkSdk();
       mtLinkSdk.init(clientId, { samlSubjectId });
 
-      await authorize(mtLinkSdk.storedOptions, {
+      const url = await authorizeUrl(mtLinkSdk.storedOptions, {
         state,
         redirectUri,
         scopes
       });
 
-      expect(open).toHaveBeenCalledTimes(1);
-      expect(open).toHaveBeenCalledWith(expect.any(String), '_self', 'noreferrer');
-      const url = open.mock.calls[0][0];
       const query = {
         client_id: clientId,
         response_type: 'code',
@@ -106,91 +95,63 @@ describe('api', () => {
 
     test('includes affiliate_code when provided', async () => {
       mockedStorage.set.mockClear();
-      open.mockClear();
 
       const state = 'state';
       const scopes = 'points_read';
-      const samlSubjectId = 'mySubject';
       const affiliateCode = 'mtb_hoge';
 
       const mtLinkSdk = new MtLinkSdk();
-      mtLinkSdk.init(clientId, { samlSubjectId });
+      mtLinkSdk.init(clientId, { redirectUri });
 
-      await authorize(mtLinkSdk.storedOptions, {
+      const url = await authorizeUrl(mtLinkSdk.storedOptions, {
         state,
         redirectUri,
         scopes,
         affiliateCode
       });
 
-      expect(open).toHaveBeenCalledTimes(1);
-      expect(open).toHaveBeenCalledWith(expect.any(String), '_self', 'noreferrer');
-      const url = open.mock.calls[0][0];
-
-      const parsed = new URL(url);
-      expect(parsed.searchParams.has('affiliate_code')).toBe(true);
+      const urlObj = new URL(url);
+      expect(urlObj.searchParams.get('affiliate_code')).toBe(affiliateCode);
     });
 
     test('does not include affiliate_code when affiliateCode is undefined', async () => {
       mockedStorage.set.mockClear();
-      open.mockClear();
 
       const state = 'state';
       const scopes = 'points_read';
-      const samlSubjectId = 'mySubject';
-      const affiliateCode = undefined;
 
       const mtLinkSdk = new MtLinkSdk();
-      mtLinkSdk.init(clientId, { samlSubjectId });
+      mtLinkSdk.init(clientId, { redirectUri });
 
-      await authorize(mtLinkSdk.storedOptions, {
+      const url = await authorizeUrl(mtLinkSdk.storedOptions, {
         state,
         redirectUri,
         scopes,
-        affiliateCode
+        affiliateCode: undefined
       });
 
-      expect(open).toHaveBeenCalledTimes(1);
-      expect(open).toHaveBeenCalledWith(expect.any(String), '_self', 'noreferrer');
-      const url = open.mock.calls[0][0];
-
-      const parsed = new URL(url);
-      expect(parsed.searchParams.has('affiliate_code')).toBe(false);
+      const urlObj = new URL(url);
+      expect(urlObj.searchParams.has('affiliate_code')).toBe(false);
     });
 
-    test('does not include affiliate_code when affiliateCode is null', async () => {
+    test('does not include affiliate_code when affiliateCode is empty string', async () => {
       mockedStorage.set.mockClear();
-      open.mockClear();
 
       const state = 'state';
       const scopes = 'points_read';
-      const samlSubjectId = 'mySubject';
-      const affiliateCode = null;
 
       const mtLinkSdk = new MtLinkSdk();
-      mtLinkSdk.init(clientId, { samlSubjectId });
+      mtLinkSdk.init(clientId, { redirectUri });
 
-      await authorize(mtLinkSdk.storedOptions, {
+      const url = await authorizeUrl(mtLinkSdk.storedOptions, {
         state,
         redirectUri,
         scopes,
-        affiliateCode
-      } as unknown as AuthorizeOptions);
+        affiliateCode: ''
+      });
 
-      expect(open).toHaveBeenCalledTimes(1);
-      expect(open).toHaveBeenCalledWith(expect.any(String), '_self', 'noreferrer');
-      const url = open.mock.calls[0][0];
-
-      const parsed = new URL(url);
-      expect(parsed.searchParams.has('affiliate_code')).toBe(false);
-    });
-
-    test('without window', async () => {
-      vi.spyOn(helper, 'hasWindow').mockReturnValue(false);
-
-      await expect(authorize(new MtLinkSdk().storedOptions)).rejects.toThrow(
-        '[mt-link-sdk] `authorize` only works in the browser.'
-      );
+      const urlObj = new URL(url);
+      expect(urlObj.searchParams.has('affiliate_code')).toBe(false);
     });
   });
 });
