+++ src/api/authorize/authorize-url.ts
@@ -1,7 +1,13 @@
-import { constructScopes, generateConfigs, mergeConfigs, generateCodeChallenge, objectToQueryString } from '../helper';
-import { MY_ACCOUNT_DOMAINS } from '../server-paths';
-import { StoredOptions, AuthorizeUrlOptions } from '../typings';
-import storage from '../storage';
+import {
+  constructScopes,
+  generateConfigs,
+  mergeConfigs,
+  generateCodeChallenge,
+  objectToQueryString
+} from '../../helper';
+import { MY_ACCOUNT_DOMAINS } from '../../server-paths';
+import { StoredOptions, AuthorizeUrlOptions } from '../../typings';
+import storage from '../../storage';
 
 export default async function authorize(
   storedOptions: StoredOptions,
