+++ src/api/authorize/authorize.test.ts
@@ -1,21 +1,24 @@
 import { MY_ACCOUNT_DOMAINS } from '../../server-paths';
-import { MtLinkSdk } from '../..';
-import authorizeUrl from '../authorize-url';
+import { AuthorizeOptions, MtLinkSdk } from '../..';
+import authorize from './authorize';
 import { generateConfigs } from '../../helper';
+import * as helper from '../../helper';
 import storage from '../../storage';
 import expectUrlToMatchWithPKCE from '../../__tests__/helper/expect-url-to-match';
 
 vi.mock('../../storage');
 
 describe('api', () => {
-  describe('authorize-url', () => {
+  describe('authorize', () => {
+    const open = (window.open = vi.fn());
+
     const mockedStorage = vi.mocked(storage);
 
     const clientId = 'clientId';
     const redirectUri = 'redirectUri';
 
     test('without calling init', async () => {
-      await expect(authorizeUrl(new MtLinkSdk().storedOptions)).rejects.toThrow(
+      await expect(authorize(new MtLinkSdk().storedOptions)).rejects.toThrow(
         '[mt-link-sdk] Make sure to call `init` before calling `authorizeUrl/authorize`.'
       );
     });
@@ -24,13 +27,14 @@ describe('api', () => {
       const mtLinkSdk = new MtLinkSdk();
       mtLinkSdk.init(clientId);
 
-      await expect(authorizeUrl(mtLinkSdk.storedOptions)).rejects.toThrow(
+      await expect(authorize(mtLinkSdk.storedOptions)).rejects.toThrow(
         '[mt-link-sdk] Missing option `redirectUri` in `authorizeUrl/authorize`, make sure to pass one via `authorizeUrl/authorize` options or `init` options.'
       );
     });
 
     test('method call without options use default init value', async () => {
       mockedStorage.set.mockClear();
+      open.mockClear();
 
       const country = 'JP';
       const scopes = 'points_read';
@@ -47,8 +51,11 @@ describe('api', () => {
         samlSubjectId
       });
 
-      const url = await authorizeUrl(mtLinkSdk.storedOptions);
+      await authorize(mtLinkSdk.storedOptions);
 
+      expect(open).toHaveBeenCalledTimes(1);
+      expect(open).toHaveBeenCalledWith(expect.any(String), '_self', 'noreferrer');
+      const url = open.mock.calls[0][0];
       const query = {
         client_id: clientId,
         cobrand_client_id: cobrandClientId,
@@ -65,6 +72,7 @@ describe('api', () => {
 
     test('with options', async () => {
       mockedStorage.set.mockClear();
+      open.mockClear();
 
       const state = 'state';
       const country = 'JP';
@@ -74,12 +82,15 @@ describe('api', () => {
       const mtLinkSdk = new MtLinkSdk();
       mtLinkSdk.init(clientId, { samlSubjectId });
 
-      const url = await authorizeUrl(mtLinkSdk.storedOptions, {
+      await authorize(mtLinkSdk.storedOptions, {
         state,
         redirectUri,
         scopes
       });
 
+      expect(open).toHaveBeenCalledTimes(1);
+      expect(open).toHaveBeenCalledWith(expect.any(String), '_self', 'noreferrer');
+      const url = open.mock.calls[0][0];
       const query = {
         client_id: clientId,
         response_type: 'code',
@@ -95,63 +106,91 @@ describe('api', () => {
 
     test('includes affiliate_code when provided', async () => {
       mockedStorage.set.mockClear();
+      open.mockClear();
 
       const state = 'state';
       const scopes = 'points_read';
+      const samlSubjectId = 'mySubject';
       const affiliateCode = 'mtb_hoge';
 
       const mtLinkSdk = new MtLinkSdk();
-      mtLinkSdk.init(clientId, { redirectUri });
+      mtLinkSdk.init(clientId, { samlSubjectId });
 
-      const url = await authorizeUrl(mtLinkSdk.storedOptions, {
+      await authorize(mtLinkSdk.storedOptions, {
         state,
         redirectUri,
         scopes,
         affiliateCode
       });
 
-      const urlObj = new URL(url);
-      expect(urlObj.searchParams.get('affiliate_code')).toBe(affiliateCode);
+      expect(open).toHaveBeenCalledTimes(1);
+      expect(open).toHaveBeenCalledWith(expect.any(String), '_self', 'noreferrer');
+      const url = open.mock.calls[0][0];
+
+      const parsed = new URL(url);
+      expect(parsed.searchParams.has('affiliate_code')).toBe(true);
     });
 
     test('does not include affiliate_code when affiliateCode is undefined', async () => {
       mockedStorage.set.mockClear();
+      open.mockClear();
 
       const state = 'state';
       const scopes = 'points_read';
+      const samlSubjectId = 'mySubject';
+      const affiliateCode = undefined;
 
       const mtLinkSdk = new MtLinkSdk();
-      mtLinkSdk.init(clientId, { redirectUri });
+      mtLinkSdk.init(clientId, { samlSubjectId });
 
-      const url = await authorizeUrl(mtLinkSdk.storedOptions, {
+      await authorize(mtLinkSdk.storedOptions, {
         state,
         redirectUri,
         scopes,
-        affiliateCode: undefined
+        affiliateCode
       });
 
-      const urlObj = new URL(url);
-      expect(urlObj.searchParams.has('affiliate_code')).toBe(false);
+      expect(open).toHaveBeenCalledTimes(1);
+      expect(open).toHaveBeenCalledWith(expect.any(String), '_self', 'noreferrer');
+      const url = open.mock.calls[0][0];
+
+      const parsed = new URL(url);
+      expect(parsed.searchParams.has('affiliate_code')).toBe(false);
     });
 
-    test('does not include affiliate_code when affiliateCode is empty string', async () => {
+    test('does not include affiliate_code when affiliateCode is null', async () => {
       mockedStorage.set.mockClear();
+      open.mockClear();
 
       const state = 'state';
       const scopes = 'points_read';
+      const samlSubjectId = 'mySubject';
+      const affiliateCode = null;
 
       const mtLinkSdk = new MtLinkSdk();
-      mtLinkSdk.init(clientId, { redirectUri });
+      mtLinkSdk.init(clientId, { samlSubjectId });
 
-      const url = await authorizeUrl(mtLinkSdk.storedOptions, {
+      await authorize(mtLinkSdk.storedOptions, {
         state,
         redirectUri,
         scopes,
-        affiliateCode: ''
-      });
+        affiliateCode
+      } as unknown as AuthorizeOptions);
 
-      const urlObj = new URL(url);
-      expect(urlObj.searchParams.has('affiliate_code')).toBe(false);
+      expect(open).toHaveBeenCalledTimes(1);
+      expect(open).toHaveBeenCalledWith(expect.any(String), '_self', 'noreferrer');
+      const url = open.mock.calls[0][0];
+
+      const parsed = new URL(url);
+      expect(parsed.searchParams.has('affiliate_code')).toBe(false);
+    });
+
+    test('without window', async () => {
+      vi.spyOn(helper, 'hasWindow').mockReturnValue(false);
+
+      await expect(authorize(new MtLinkSdk().storedOptions)).rejects.toThrow(
+        '[mt-link-sdk] `authorize` only works in the browser.'
+      );
     });
   });
 });
