+++ src/api/authorize/authorize.ts
@@ -1,5 +1,5 @@
-import { getIsTabValue, hasWindow, openWindow } from '../helper';
-import { StoredOptions, AuthorizeOptions } from '../typings';
+import { getIsTabValue, hasWindow, openWindow } from '../../helper';
+import { StoredOptions, AuthorizeOptions } from '../../typings';
 import authorizeUrl from './authorize-url';
 
 export default async function authorize(storedOptions: StoredOptions, options: AuthorizeOptions = {}): Promise<void> {
