+++ src/api/exchange-token/exchange-token.ts
@@ -1,7 +1,7 @@
-import { generateSdkHeaderInfo, hasWindow, queryStringToObject } from '../helper';
-import { MY_ACCOUNT_DOMAINS } from '../server-paths';
-import { StoredOptions, ExchangeTokenOptions, Token } from '../typings';
-import storage from '../storage';
+import { generateSdkHeaderInfo, hasWindow, queryStringToObject } from '../../helper';
+import { MY_ACCOUNT_DOMAINS } from '../../server-paths';
+import { StoredOptions, ExchangeTokenOptions, Token } from '../../typings';
+import storage from '../../storage';
 
 function getCode(): string | undefined {
   // not available in node environment
