+++ src/api/logout/logout.test.ts
@@ -1,6 +1,6 @@
 import { MY_ACCOUNT_DOMAINS } from '../../server-paths';
 import { MtLinkSdk } from '../..';
-import logout from '../logout';
+import logout from './logout';
 import { generateConfigs, objectToQueryString } from '../../helper';
 import * as helper from '../../helper';
